HPCC-34366 Refactor the lz4/zstd code - remove unused functions
