+++ system/jlib/jzstd.hpp
@@ -1,6 +1,6 @@
 /*##############################################################################
 
-    HPCC SYSTEMS software Copyright (C) 2015 HPCC Systems.
+    HPCC SYSTEMS software Copyright (C) 2025 HPCC Systems.
 
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
@@ -15,16 +15,12 @@
     limitations under the License.
 ############################################################################## */
 
-#ifndef JLZ4_INCL
-#define JLZ4_INCL
+#ifndef JZSTD_INCL
+#define JZSTD_INCL
 
 #include "jlzw.hpp"
 
-#define LZ4COMPRESSEDFILEBLOCKSIZE (0x100000)
-
-extern jlib_decl ICompressor *createLZ4Compressor(const char * options, bool hc=false);
-extern jlib_decl IExpander   *createLZ4Expander();
-extern jlib_decl ICompressor *createLZ4StreamCompressor(const char * options, bool hc=false);
-extern jlib_decl IExpander   *createLZ4StreamExpander();
+extern jlib_decl ICompressor *createZStdStreamCompressor(const char * options);
+extern jlib_decl IExpander   *createZStdStreamExpander();
 
 #endif
